+++ src/org/meresco/lucene/search/InterpolateEpsilon.java
@@ -28,7 +28,7 @@ package org.meresco.lucene.search;
 
 
 public class InterpolateEpsilon {
-    public double interpolateEpsilon(int totalHits, int sliceSize, double clusteringEps, int clusterMoreRecords) {
+    public double interpolateEpsilon(long totalHits, int sliceSize, double clusteringEps, int clusterMoreRecords) {
         double eps = clusteringEps * (totalHits - sliceSize) / clusterMoreRecords;
         return Math.max(Math.min(eps, clusteringEps), 0.0);
     }
